+++ docs/conf.py
@@ -13,9 +13,9 @@ copyright = '2023, MatrixEditor'
 author = 'MatrixEditor'
 
 # The short X.Y version.
-version = '0.0'
+version = '0.1'
 # The full version, including alpha/beta/rc tags.
-release = '0.0.1'
+release = '0.1.0'
 
 # If extensions (or modules to document with autodoc) are in another directory,
 # add these directories to sys.path here. If the directory is relative to the