+++ examples/example.py
@@ -5,7 +5,6 @@ writer = SmaliWriter(reader)
 
 with open('example.smali', 'r', encoding='utf-8') as fp:
     source = fp.read()
-    
 
 reader.visit(source, writer)
 print(writer.code)