+++ examples/shell_example.py
@@ -1,0 +1,8 @@
+from smali.shell import ISmaliShell, start_cli
+
+# Just start the shell in your code or
+# call it from the CLI with 'python3 -m smali.shell'
+ISmaliShell().cmdloop()
+
+# Or just call start CLI
+start_cli()