+++ examples/vm_example.py
@@ -2,7 +2,7 @@ from smali.bridge import SmaliVM, SmaliObject
 
 with open('example.smali', 'r', encoding='utf-8') as fp:
     source = fp.read()
-    
+
 vm = SmaliVM()
 smali_class = vm.classloader.load_class(source, init=False)
 smali_class.clinit()