+++ smali/__init__.py
@@ -24,3 +24,5 @@ from smali.reader import *
 from smali.writer import *
 
 SmaliValue = smali_value
+
+VERSION = '0.2.0'