+++ docs/changelog.rst
@@ -4,6 +4,15 @@
 Changelog
 *********
 
+.. _release-0.1.0:
+
+0.1.0
+=====
+
+* ISmali (interactive Smali shell) pre-release
+* Implementation of almost half of all Smali-opcodes
+* Fixed an error of ``SmaliValueProxy`` that caused exceptions on operations with an object of the same class
+* Multiple bug fixes in the process of class definition (import)
 
 .. _release-1.0.0: